Se ha agregado la clase CostoMinimos

+++ MetodoDeTransporte/MetodoDeTransporte/Datos.cs
@@ -13,11 +13,11 @@ namespace MetodoDeTransporte
         public static double resultadoGlobal = 0;
 
         //atributos de la clase
-        private int cantidad = 0;
+        private double cantidad = 0;
         private double precio = 0; 
         private bool llena = false;
         //setters y getters de la clase
-        public int Cantidad { set { this.cantidad = value; } get { return this.cantidad; } }
+        public double Cantidad { set { this.cantidad = value; } get { return this.cantidad; } }
         public double Precio { set { this.precio = value; } get { return this.precio; } }
         public bool Llena { set { this.llena = value; } get { return this.llena; } }
